+++ src/cloudant/_messages.py
@@ -73,7 +73,7 @@ CLIENT = {
     101: 'Value must be set to a Database object. Found type: {0}',
     102: 'You must provide a url or an account.',
     404: 'Database {0} does not exist. Verify that the client is valid and try again.',
-    409: 'Database {0} already exists.'
+    412: 'Database {0} already exists.'
 }
 
 DATABASE = {
@@ -81,7 +81,8 @@ DATABASE = {
     101: 'Unexpected index type. Found: {0}',
     400: 'Invalid database name during creation. Found: {0}',
     401: 'Unauthorized to create database {0}',
-    409: 'Document with id {0} already exists.'
+    409: 'Document with id {0} already exists.',
+    412: 'Database {0} already exists.'
 }
 
 DESIGN_DOCUMENT = {